Modified Mongo class and added AdminMongo class.
Modified the mongo class to make it compatible with the soon to come startup class. Did this by just having the name, id and score passed directly to the class. Also disabled showing the default pymongo id, as I decided to use the uuid package to generate unique ids. Along with this. I created the AdminMongo class, which houses some of the features which would break the Mongo object if they were used and then other functions were attempted. (Delete database, than try to store to collection.)